requirements.txt
absl-py==0.2.0
alabaster==0.7.11
astor==0.6.2
Babel==2.6.0
backcall==0.1.0
bleach==1.5.0
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
Django==2.0.6
django-cors-headers==2.3.0
django-des==2.4.1
django-solo==1.1.3
djangorestframework==3.8.2
djmail==1.1.0
docutils==0.14
entrypoints==0.2.3
gast==0.2.0
grpcio==1.11.0
h5py==2.7.1
html5lib==0.9999999
idna==2.7
imagesize==1.0.0
ipdb==0.11
ipykernel==4.8.2
ipyparallel==6.2.1
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
notebook==5.4.1
numpy==1.14.2
opencv-python==3.4.1.15
packaging==17.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
psycopg2==2.7.5
ptyprocess==0.5.2
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
pyparsing==2.2.0
python-apt==1.1.0b1+ubuntu0.16.4.1
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.13
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.19.1
scikit-learn==0.19.1
scipy==1.0.1
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.7.5
sphinxcontrib-websupport==1.1.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
